+++ src/main/java/agents/flexibility/dynamicProgramming/assessment/AssessmentFunction.java
@@ -58,4 +58,11 @@ public interface AssessmentFunction {
 	default double getMultiplier() {
 		return 1.0;
 	}
+
+	/** Get the electricity price predicted for the given amount of dispatched energy at the specified time
+	 * 
+	 * @param time at which to calculate the predicted electricity price
+	 * @param externalEnergyDeltaInMWH charging if positive, discharging when negative
+	 * @return predicted electricity price */
+	double getElectricityPriceAt(TimeStamp time, double externalEnergyDeltaInMWH);
 }